Upgrade several build dependencies. (#11812)
* Upgrade several build dependencies.

* Update error prone rules (those are off but they do trigger warnings/ errors)

* A few corrections I made before I turned off new warnings. Let's do nother issue to fix them.

+++ lucene/core/src/java/org/apache/lucene/store/ByteBuffersDataOutput.java
@@ -459,7 +459,7 @@ public final class ByteBuffersDataOutput extends DataOutput implements Accountab
     // any special distinction for direct memory buffers.
     assert ramBytesUsed
         == blocks.stream().mapToLong(ByteBuffer::capacity).sum()
-            + blocks.size() * RamUsageEstimator.NUM_BYTES_OBJECT_REF;
+            + (long) blocks.size() * RamUsageEstimator.NUM_BYTES_OBJECT_REF;
     return ramBytesUsed;
   }
